[Roopa] Added: FeedBack API's are Done

+++ BookStore/CommonLayer/User/BookModel.cs
@@ -14,7 +14,7 @@ namespace CommonLayer.User
         public string BookDescription { get; set; }
 
         public string Image { get; set; }
-        public float Rating { get; set; }
+        public Double Rating { get; set; }
         public int Reviewer { get; set; }
         public int BookCount { get; set; }
     }
